Fix test_list_pkgs_homebrew_cask_pakages
